Test Sorting Package: * Create test_method_generator_sort * Move test/sorting/__init__ to test/changelist_sort/sorting/__init__ * Move test/sorting/test_developer_sort to test/changelist_sort/sorting/test_developer_sort * Move test/sorting/test_file_sort to test/changelist_sort/sorting/test_file_sort * Move test/sorting/test_list_sort to test/changelist_sort/sorting/test_list_sort * Move test/sorting/test_method_map_sort to test/changelist_sort/sorting/test_method_map_sort * Move test/sorting/test_method_sort to test/changelist_sort/sorting/test_method_sort * Move test/sorting/test_module_sort to test/changelist_sort/sorting/test_module_sort * Move test/sorting/test_sorting_changelist to test/changelist_sort/sorting/test_sorting_changelist * Move test/sorting/test_sorting_file_pattern to test/changelist_sort/sorting/test_sorting_file_pattern * Move test/sorting/test_source_set_sort to test/changelist_sort/sorting/test_source_set_sort * Move test/sorting/test_string_operations to test/changelist_sort/sorting/test_string_operations
